+++ devicelayer/src/DevicePcie.h
@@ -21,14 +21,13 @@ public:
   ~DevicePcie();
 
   // IDeviceAsync
-  bool sendCommandMasterMinion(int device, int sqIdx, std::byte* command, size_t commandSize, bool isDma,
-                               bool isHighPriority) override;
+  bool sendCommandMasterMinion(int device, int sqIdx, std::byte* command, size_t commandSize, CmdFlagMM flags) override;
   void setSqThresholdMasterMinion(int device, int sqIdx, uint32_t bytesNeeded) override;
   void waitForEpollEventsMasterMinion(int device, uint64_t& sq_bitmap, bool& cq_available,
                                       std::chrono::milliseconds timeout = std::chrono::seconds(10)) override;
   bool receiveResponseMasterMinion(int device, std::vector<std::byte>& response) override;
 
-  bool sendCommandServiceProcessor(int device, std::byte* command, size_t commandSize, bool isMmReset = false) override;
+  bool sendCommandServiceProcessor(int device, std::byte* command, size_t commandSize, CmdFlagSP flags) override;
   void setSqThresholdServiceProcessor(int device, uint32_t bytesNeeded) override;
   void waitForEpollEventsServiceProcessor(int device, bool& sq_available, bool& cq_available,
                                           std::chrono::milliseconds timeout = std::chrono::seconds(10)) override;
@@ -71,6 +70,9 @@ private:
     int fdMgmt_;
     int epFdMgmt_;
   };
+
+  void setupDeviceInfo(int device, DevInfo& deviceInfo, std::chrono::seconds timeout = std::chrono::seconds(30)) const;
+  void teardownDeviceInfo(const DevInfo& deviceInfo) const;
 
   std::unordered_map<void*, size_t> dmaBuffers_;
   std::vector<DevInfo> devices_;